Refactor: create Excel Spreadsheet class apart from model spreadsheet class
