+++ src/main/java/com/tiagoamp/sjc/model/input/InRow.java
@@ -7,7 +7,7 @@ public class InRow {
 	private Integer qtdHoraExtra;
 	private Integer qtdAdicionalNoturno;
 	private Integer qtdPlantoesExtra;
-	private String[] dtPlantoesExtras = new String[5];  //FIXME should be LocalDate array !! 
+	private String[] dtPlantoesExtras = new String[5];  
 	
 	
 	@Override
