+++ src/main/java/com/tiagoamp/sjc/service/SpreadsheetServices.java
@@ -13,11 +13,13 @@ import org.springframework.stereotype.Service;
 
 import com.itextpdf.text.DocumentException;
 import com.tiagoamp.sjc.model.input.InputSpreadsheet;
+import com.tiagoamp.sjc.model.input.InputExcelSpreadsheet;
 import com.tiagoamp.sjc.model.output.OutputSpreadsheet;
+import com.tiagoamp.sjc.model.output.OutputExcelSpreadsheet;
 
 @Service
-public class SjcServicesFacade {
-
+public class SpreadsheetServices {
+	
 	/**
 	 * Load data from a spreadsheet of a given file path.
 	 * 
@@ -26,10 +28,9 @@ public class SjcServicesFacade {
 	 * @throws IOException
 	 */
 	public InputSpreadsheet loadInputSpreadsheet(Path filepath) throws IOException {
-		if (Files.notExists(filepath)) throw new IllegalArgumentException("Diretório inexistente!");
-		InputSpreadsheet spreadsheet = new InputSpreadsheet();
-		spreadsheet.loadFromFile(filepath);	
-		return spreadsheet;
+		if (Files.notExists(filepath)) throw new IllegalArgumentException("Arquivo inexistente!");
+		InputExcelSpreadsheet excelSheet = new InputExcelSpreadsheet(filepath);
+		return excelSheet.toInputSpreadsheet();
 	}
 	
 	public List<InputSpreadsheet> loadInputSpreadsheetsFromDirectory(Path directory) throws IOException {
@@ -37,22 +38,20 @@ public class SjcServicesFacade {
 		List<InputSpreadsheet> inputList = new ArrayList<>();
 		DirectoryStream<Path> stream = Files.newDirectoryStream(directory);
 		for (Path file : stream) {
-			InputSpreadsheet spreadsheet = new InputSpreadsheet(); 
-			spreadsheet.loadFromFile(file);	
+			InputExcelSpreadsheet excelSheet = new InputExcelSpreadsheet(file);
+			InputSpreadsheet spreadsheet = excelSheet.toInputSpreadsheet();
 			inputList.add(spreadsheet);				
 		}		
 		return inputList;
 	}
 	
-	public OutputSpreadsheet generateOutputSpreadSheet(List<InputSpreadsheet> list) {
-		OutputSpreadsheet outputSpreadSheet = new OutputSpreadsheet();
-		outputSpreadSheet.loadDataFromInputSpreadSheets(list);
-		return outputSpreadSheet;
+	public OutputSpreadsheet generateOutputSpreadSheet(List<InputSpreadsheet> spreadsheetsList) {
+		return outSpreadsheetFactory.loadDataFromInputSpreadsheets(spreadsheetsList);
 	}
 	
-	public void generateOuputSpreadsheetFile(Path outputFile, OutputSpreadsheet outputSpreadsheet, Path templateFile) throws IOException {
-		if (templateFile == null) templateFile = Paths.get("resources","template_output.xlsx");		
-		outputSpreadsheet.generateOuputSpreadsheetFile(outputFile, templateFile);
+	public void generateOuputSpreadsheetFile(Path outputFile, OutputSpreadsheet outputSpreadsheet) throws IOException {
+		//if (templateFile == null) templateFile = Paths.get("resources","template_output.xlsx");		
+		outputSpreadsheet.generateOuputSpreadsheetFile(outputFile);
 	}
 	
 	public void generateOutputMessagesFile(Path outputFile, OutputSpreadsheet outputSpreadsheet) throws FileNotFoundException, DocumentException  {
