+++ src/test/java/com/tiagoamp/sjc/model/input/InputSpreadsheetTest.java
@@ -1,62 +1,77 @@
 package com.tiagoamp.sjc.model.input;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.time.YearMonth;
+import java.util.Arrays;
+import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 import com.tiagoamp.sjc.model.MessageType;
+import com.tiagoamp.sjc.model.ProcessingMessage;
+import com.tiagoamp.sjc.model.SjcGeneralCode;
 
 public class InputSpreadsheetTest {
 	
-	private InputSpreadsheet inputSpreadsheet;
+	private InputSpreadsheet spreadsheet;
 
 	@Before
 	public void setUp() throws Exception {
-		inputSpreadsheet = new InputSpreadsheet();
+		spreadsheet = new InputSpreadsheet();
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		inputSpreadsheet = null;
+		spreadsheet = null;
 	}
 
-	
 	@Test
-	public void testLoadFromFile_directory_shouldReturnErrorProcessingMessage() throws IOException {
-		Path directoryFile = Paths.get("testfiles","entrada","dummyDirectory");		
-		inputSpreadsheet.loadFromFile(directoryFile);
-		assertTrue("Processing messages must not be empty", !inputSpreadsheet.getMessages().isEmpty());
-		assertEquals("Must have directory error message", MessageType.ERROR, inputSpreadsheet.getMessages().get(0).getType());
-		assertTrue("Must have directory error message", inputSpreadsheet.getMessages().get(0).getText().contains("diretório"));
-	}
-	
-	@Test
-	public void testLoadFromFile_notXLSXfile_shouldReturnErrorProcessingMessage() throws IOException {
-		Path dummyFile = Paths.get("testfiles","entrada","dummyFile.dmp");
-		inputSpreadsheet.loadFromFile(dummyFile);
-		assertTrue("Processing messages must not be empty", !inputSpreadsheet.getMessages().isEmpty());
-		assertEquals("Must have error message", MessageType.ERROR, inputSpreadsheet.getMessages().get(0).getType());
-		assertTrue("Must have xlsx error message", inputSpreadsheet.getMessages().get(0).getText().contains("xlsx"));
-	}
-	
-	@Test
-	public void testLoadFromFile_emptySheet_shouldReturnErrorProcessingMessage() throws IOException {
-		Path inputFile = Paths.get("testfiles","entrada","blank_sheet.xlsx");
-		inputSpreadsheet.loadFromFile(inputFile);
-		assertTrue("Processing messages must not be empty", !inputSpreadsheet.getMessages().isEmpty());
-		assertEquals("Must have 3 errors message", 3, inputSpreadsheet.getMessages().size());
-		assertTrue("Must have sheet not found error message (1st)", inputSpreadsheet.getMessages().get(0).getText().contains("não encontrada"));
-		assertTrue("Must have sheet not found error message (2nd)", inputSpreadsheet.getMessages().get(1).getText().contains("não encontrada"));
-		assertTrue("Must have 'lotacao' not found error message (3rd)", inputSpreadsheet.getMessages().get(2).getText().contains("Não foi identificado"));
-		assertEquals("'Lotacao' must be setted to expected", "!NOME DA UNIDADE NÃO IDENTIFICADO NA PLANILHA!", inputSpreadsheet.getLotacao());
+	public void getMessages_hasSpreadsheetAndInSheetMessages_shouldReturnInternalAndOuterMessages() {
+		// given
+		ProcessingMessage spreadsheetMessage = new ProcessingMessage(MessageType.ERROR,"MSG00");
+		spreadsheet.getMessages().add(spreadsheetMessage);
+		List<ProcessingMessage> sheetMessages = Arrays.asList(new ProcessingMessage(MessageType.ERROR,"MSG01"),new ProcessingMessage(MessageType.ALERT,"MSG02"));
+		InSheet sheet =  new InSheet(SjcGeneralCode.ADMINISTRATIVO, YearMonth.now());
+		sheet.setMessages(sheetMessages);
+		spreadsheet.getSheets().put(sheet.getCode(), sheet);		
 		
+		// when
+		List<ProcessingMessage> allMessages = spreadsheet.getMessages();
+		
+		// then
+		assertEquals(3, allMessages.size());
 	}
 	
+	@Test
+	public void getMessages_hasOnlySpreadsheetMessages_shouldReturnOuterMessages() {
+		// given
+		ProcessingMessage spreadsheetMessage = new ProcessingMessage(MessageType.ERROR,"MSG00");
+		spreadsheet.getMessages().add(spreadsheetMessage);
+		InSheet sheet =  new InSheet(SjcGeneralCode.ADMINISTRATIVO, YearMonth.now());
+		spreadsheet.getSheets().put(sheet.getCode(), sheet);		
+		
+		// when
+		List<ProcessingMessage> allMessages = spreadsheet.getMessages();
+		
+		// then
+		assertEquals(1, allMessages.size());
+	}
+	
+	@Test
+	public void getMessages_hasOnlyInSheetMessages_shouldReturnInternalMessages() {
+		// given
+		List<ProcessingMessage> sheetMessages = Arrays.asList(new ProcessingMessage(MessageType.ERROR,"MSG01"),new ProcessingMessage(MessageType.ALERT,"MSG02"));
+		InSheet sheet =  new InSheet(SjcGeneralCode.ADMINISTRATIVO, YearMonth.now());
+		sheet.setMessages(sheetMessages);
+		spreadsheet.getSheets().put(sheet.getCode(), sheet);		
+		
+		// when
+		List<ProcessingMessage> allMessages = spreadsheet.getMessages();
+		
+		// then
+		assertEquals(2, allMessages.size());
+	}
 }
