+++ src/test/java/com/tiagoamp/sjc/service/SpreadsheetServicesTest.java
@@ -16,9 +16,9 @@ import org.junit.Test;
 import com.tiagoamp.sjc.model.input.InputSpreadsheet;
 import com.tiagoamp.sjc.model.output.OutputSpreadsheet;
 
-public class SjcServicesFacadeTest {
+public class SpreadsheetServicesTest {
 	
-	private SjcServicesFacade facade;
+	private SpreadsheetServices service;
 	private static Path inputDir = Paths.get("testfiles", "entrada");
 	private static Path outputDir = Paths.get("testfiles", "saida");
 
@@ -31,40 +31,40 @@ public class SjcServicesFacadeTest {
 	
 	@Before
 	public void setUp() throws Exception {
-		facade = new SjcServicesFacade();		
+		service = new SpreadsheetServices();		
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		facade = null;
+		service = null;
 	}
 
 	@Test
 	public void testLoadInputSpreadSheet_shouldLoadFile() throws IOException {
-		InputSpreadsheet result = facade.loadInputSpreadsheet(inputDir.resolve("template_input.xlsx"));
+		InputSpreadsheet result = service.loadInputSpreadsheet(inputDir.resolve("template_input.xlsx"));
 		assertNotNull("Must generate the input spreadsheet.", result);
 	}
 	
 	@Test
 	public void testLoadInputSpreadsheetsFromDirectory_shouldLoadFiles() throws IOException {
-		List<InputSpreadsheet> result = facade.loadInputSpreadsheetsFromDirectory(inputDir);
+		List<InputSpreadsheet> result = service.loadInputSpreadsheetsFromDirectory(inputDir);
 		assertNotNull("Must generate list of input spreadsheets.", result);
 	}
 
 	@Test
 	public void testGenerateOutputSpreadSheet_shouldGenerateValidOutput() throws IOException {
-		List<InputSpreadsheet> inputlist = facade.loadInputSpreadsheetsFromDirectory(inputDir);
-		OutputSpreadsheet result = facade.generateOutputSpreadSheet(inputlist);
+		List<InputSpreadsheet> inputlist = service.loadInputSpreadsheetsFromDirectory(inputDir);
+		OutputSpreadsheet result = service.generateOutputSpreadSheet(inputlist);
 		assertNotNull("Must generate output spreadsheets.", result);
 		
 	}
 
 	@Test
 	public void testGenerateOuputSpreadsheetFile_shouldGenerateValidOutput() throws IOException {
-		List<InputSpreadsheet> inputlist = facade.loadInputSpreadsheetsFromDirectory(inputDir);
-		OutputSpreadsheet outspreadsheet = facade.generateOutputSpreadSheet(inputlist);
+		List<InputSpreadsheet> inputlist = service.loadInputSpreadsheetsFromDirectory(inputDir);
+		OutputSpreadsheet outspreadsheet = service.generateOutputSpreadSheet(inputlist);
 		Path outputfile = outputDir.resolve("testOutFromFacadeTest.xlsx");
-		facade.generateOuputSpreadsheetFile(outputfile, outspreadsheet, null);
+		service.generateOuputSpreadsheetFile(outputfile, outspreadsheet, null);
 		assertTrue("Must generate output spreadsheet file.", Files.exists(outputfile));
 	}
